* Release version 0.33 * Fix ancient int/long truncation bug in LengthInBlocks * Fix small bug in file count/size reporting during Update * Change copyright year to 2013

+++ disParityLib/Parity.cs
@@ -102,7 +102,7 @@ namespace disParity
     public static UInt32 LengthInBlocks(long lengthInBytes)
     {
       UInt32 result = (UInt32)(lengthInBytes / BLOCK_SIZE);
-      if (result * BLOCK_SIZE < lengthInBytes)
+      if ((long)result * BLOCK_SIZE < lengthInBytes)
         result++;
       return result;
     }
